fix: Keep solving even through intermittent network errors
During the past few days, Leetcode has been suffering from a high rate of
errors. Previously, these indicated something went seriously wrong with the
solution, and so it was flagged as ERROR and we proceeded with the next one.
Today, all the problems ended up flagged because of that, and then flagged
again when manually retried.

The error now must be persistent, intermittent errors will just be retried.