Fix: If necessary, switch to new web design before attempting to solve a problem
When opening a new browser to solve a new problem, sometimes we end up with the
old design. This is possibly due to the previous problem solver crashing before
it had a chance to switch back to the new design, or possibly due to some
cookie handling issues between us and the Leetcode website.

Which would lead to Brute Lee being unable to solve any further problems, until
it was run in headful mode and the switch to the new design was performed
manually.

With this update, we proactively check if we're on the new design, and if we're
not, we switch to it before attempting to solve a problem.