Start on the comprehensive compression approach
We start witht he substring compression which will be useful in problem *609.
Find Duplicate File in System*.

The key insight here is that in addition to the 100kB submission, we can use,
for free, the information in the test case text.  In this case, we use the test
case text as a code book, and instead of storing text chunks in our compression
dictionary, we only store indirect references to the text chunks in the code
book.

Eventually, the Compressor will test multiple chains of compression functions, and output a combination of the compressed text and the corresponding decompression function chain, combined in self-decompressing code. In the case of problem 609, the winning combination will surely be:

```
var buffer = JSON.parse(LZMA.decompress(substringDecompress(...)))
```