fixes: red dots for unsuccessful submissions are displaying
Probably was due to the database schema having been updated, so the report is
using the new schema while the logger is an instance that has been running
since before the change; nevertheless, setting the default Y value should help
us for situations where the test suit size is unknown, i.e. all the problems
where we the solver cannot even solve one. Since these are not captured in the
coverage report, this fix enables us to see them for the first time.