fixes: binary tree - number of passed tests is not increasing
LeetCode's serialization of binary trees is sparse: descendants of null nodes
are omitted. Since our implementation of this is buggy, we use the internal
deserializer.

The nice thing about this is that their implementation is available for all the
data structures, so we may be able to use it for other things as well.

The not-so-nice thing is that the implementation is a black box, and other
languages do not have such nice introspection, so this implementation won't
translate.